+++ packages/create-moralis-dapp/.eslintrc.js
@@ -2,4 +2,7 @@ module.exports = {
   extends: ['@moralisweb3'],
   plugins: ['jest'],
   ignorePatterns: ['**/lib/**/*', '**/*.test.ts', '**/dist/**/*', '**/build/**/*', '**/generated/**/*'],
+  rules: {
+    'no-console': 'off',
+  },
 };