+++ packages/next/src/hooks/resolvers/index.ts
@@ -1,0 +1,2 @@
+export { default as useResolver } from './useResolver';
+export { default as useResolverPaginated } from './useResolverPaginated';