+++ packages/streams/src/methods/addAddress.ts
@@ -2,7 +2,11 @@ import Core from '@moralisweb3/common-core';
 import { OperationResolver } from '@moralisweb3/api-utils';
 import { StreamNetwork } from '../utils/StreamNetwork';
 import { IncorrectNetworkError } from '../utils/IncorrectNetworkError';
-import { addAddressEvmOperation, AddAddressEvmRequest } from '@moralisweb3/common-streams-utils';
+import {
+  addAddressEvmOperation,
+  AddAddressEvmRequest,
+  AddAddressEvmResponseAdapter,
+} from '@moralisweb3/common-streams-utils';
 
 export interface AddAddressEvmOptions extends AddAddressEvmRequest {
   networkType?: 'evm';
@@ -13,7 +17,7 @@ export type AddAddressOptions = AddAddressEvmOptions;
 export const makeAddAddress = (core: Core, baseUrl: string) => {
   const evmFetcher = new OperationResolver(addAddressEvmOperation, baseUrl, core).fetch;
 
-  return ({ networkType, ...options }: AddAddressOptions) => {
+  return ({ networkType, ...options }: AddAddressOptions): Promise<AddAddressEvmResponseAdapter> => {
     switch (networkType) {
       case StreamNetwork.EVM:
         return evmFetcher({ ...options });