+++ packages/typedocReadmeTheme/src/theme.ts
@@ -3,12 +3,9 @@ import { MarkdownTheme } from 'typedoc-plugin-markdown';
 import { DeclarationReflection, PageEvent, Reflection } from 'typedoc';
 import { getPageTitle } from 'typedoc-plugin-markdown/dist/utils/front-matter';
 
-// reference to https://docs.moralis.io/docs/nodejs-technical-references
-const PARENT_ID = '6306b23a7acda000913f9ed7';
-// id for the technical references
-const CATEGORY_ID = '630696f4ab47c500b41899b7';
-
 export class ReadmeTheme extends MarkdownTheme {
+  publicPath = '';
+
   formatContents(contents: string): string {
     return `${contents
       .replace(/[\r\n]{3,}/g, '\n\n')
@@ -16,15 +13,14 @@ export class ReadmeTheme extends MarkdownTheme {
       .replace(/^\s+|\s+$/g, '')}\n`;
   }
 
-  makeFrontMatter(page: PageEvent<Reflection>): string {
-    return `---
-title: "${this.getTitle(page)}"
-category: ${CATEGORY_ID}
-parentDoc: ${PARENT_ID}
-hidden: false
+  getFrontMatter(page: PageEvent<any>) {
+    return `\
+---
+slug: "${this.getSlug(page)}"
 ---
 `;
   }
+
   getTitle(page: PageEvent<Reflection>) {
     if (page.url === this.entryDocument) {
       return page.url === page.project.url ? this.indexTitle || page.model.name : 'Readme';
@@ -35,12 +31,12 @@ hidden: false
     return getPageTitle(page as PageEvent<unknown>);
   }
 
-  getRelativeUrl(url: string) {
-    return encodeURI(url.replace('.md', ''));
-  }
-
   toUrl(mapping: any, reflection: DeclarationReflection) {
-    return `${reflection.getFullName().replace(/[/.]/g, '-').replace('@', '').toLowerCase()}.md`;
+    let url = `/${reflection.getFullName().replace(/[.]/g, '/').replace('@', '').toLowerCase()}.md`;
+    if (reflection.kindString === 'Module') {
+      url = url.replace('.md', '/index.md');
+    }
+    return url;
   }
 
   getSlug(page: PageEvent<any>) {
@@ -48,6 +44,6 @@ hidden: false
   }
 
   render(page: PageEvent<Reflection>): string {
-    return `${this.makeFrontMatter(page)}${this.formatContents(page.template(page) as string)}`;
+    return `${this.getFrontMatter(page)}${this.formatContents(page.template(page) as string)}`;
   }
 }