+++ packages/testUtils/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @moralisweb3/test-utils
 
+## 2.9.0
+
 ## 2.8.2
 
 ## 2.8.1